fix nn_reg type error with boolean md enrichment

+++ ritme/feature_space/_process_train.py
@@ -54,7 +54,6 @@ def process_train(config, train_val, target, host_id, tax, seed_data):
 
     # SPLIT
     train, val = _split_data_grouped(train_val_te, host_id, 0.8, seed_data)
-    X_train, y_train = train[ft_ls_used], train[target].astype(float)
-    X_val, y_val = val[ft_ls_used], val[target].astype(float)
-
+    X_train, y_train = train[ft_ls_used].astype(float), train[target].astype(float)
+    X_val, y_val = val[ft_ls_used].astype(float), val[target].astype(float)
     return (X_train.values, y_train.values, X_val.values, y_val.values)
